DIV-4962 - expose the endpoint to generate and return the PDF binary
To be able to attach it to an email without uploading it anywhere
